revisione struttura file
